moved some market db stuff
