+++ app/src/integration/java/com/internetEnemies/combatCritters/MarketCycleIntegrationTest.java
@@ -12,7 +12,7 @@ import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
 import com.internetEnemies.combatCritters.data.market.IMarketDB;
 import com.internetEnemies.combatCritters.data.hsqldb.CardInventoryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.CurrencyInventoryHSQLDB;
-import com.internetEnemies.combatCritters.data.hsqldb.MarketRegistryHSQLDB;
+import com.internetEnemies.combatCritters.data.market.MarketRegistryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.PackInventoryHSQLDB;
 
 import com.internetEnemies.combatCritters.data.hsqldb.TransactionRegistryHSQLDB;
